requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-jazzmin==3.0.1
djangorestframework==3.16.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
kombu==5.5.4
ngrok==1.4.0
openai==2.6.1
outcome==1.3.0.post0
packaging==25.0
playwright==1.55.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyee==13.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==7.0.0
requests==2.32.5
selenium==4.37.0
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.2.0
